Function update_page_fields split in module field_update
