+++ PyPDF2_Fields/__init__.py
@@ -1,3 +1,6 @@
+from .field_update import\
+	update_page_fields
+
 from .field_types import\
 	PdfFieldType,\
 	get_field_type
@@ -5,7 +8,6 @@ from .field_types import\
 from .misc import\
 	make_writer_from_reader,\
 	pair_fields_name_and_val,\
-	set_need_appearances,\
-	update_page_fields
+	set_need_appearances
 
 from .radio_btn_group import RadioBtnGroup
